Fix 32bit build
There's a BDecimalSpinner::SetValue(double) and SetValue(int32).
On 32bit we need to explicitely cast to int32.

+++ source/ffgui-window.cpp
@@ -275,7 +275,7 @@ ffguiwin::ffguiwin(BRect r, const char* name, window_type type, ulong mode)
 
 	// set the initial values
 	fVideoBitrateSpinner->SetValue(1000);
-	fFramerate->SetValue(30);
+	fFramerate->SetValue(int32(30));
 	fXres->SetValue(1280);
 	fYres->SetValue(720);
 	fChannelCount->SetValue(2);
@@ -1183,7 +1183,7 @@ ffguiwin::SetDefaults()
 {
 	// set the initial values
 	fVideoBitrateSpinner->SetValue(1000);
-	fFramerate->SetValue(30);
+	fFramerate->SetValue(int32(30));
 	fXres->SetValue(1280);
 	fYres->SetValue(720);
